+++ packages/std/src/object/pickBy.ts
@@ -8,11 +8,17 @@ import { createMergeTarget } from './merge';
  */
 export default function pickBy<T extends object>(
   obj: T,
-  predicate: (value: T[keyof T], key: string) => boolean,
+  predicate: (value: T[keyof T], key: string | symbol) => boolean,
 ): Partial<T> {
   const result: any = createMergeTarget(obj);
+  const keys = [
+    ...Object.keys(obj),
+    ...Object.getOwnPropertySymbols(obj).filter((sym) =>
+      Object.prototype.propertyIsEnumerable.call(obj, sym),
+    ),
+  ] as (string | symbol)[];
 
-  for (const key of Object.keys(obj)) {
+  for (const key of keys) {
     const val = obj[key as keyof T];
     if (predicate(val, key)) {
       result[key] = val;